LayoutTests/imported/w3c/web-platform-tests/svg/geometry/svg-image-intrinsic-size-with-cssstyle-auto-dynamic-image-change-expected.txt


FAIL Test <svg:image>'s sizing with css size as auto, with dynamic image change assert_equals: expected 256 but got 128
